refactor: app wave 9 slice B densify and budget compliance (#4117)
Consolidate AppEventHandlerScope de-part libraries (subscriptions,
lifecycle dialog builders, train-order helpers) to stay within the
60_550 main-line budget. Add game_service cluster to the no-part
grandfather until slice A library migration lands.
