--- app/lib/core/services/app_event_handler/app_event_handler_scope_log.dart
@@ -1,4 +1,0 @@
-import 'package:colonizethis_app/package_logger.dart';
-
-/// Shared logger for [AppEventHandlerScope] session wiring.
-final appEventHandlerScopeLog = packageLogger('event');
