+++ tool/app_no_part_directives_grandfather.dart
@@ -8,6 +8,14 @@ const List<String> appNoPartDirectivesGrandfatheredForTests = <String>[
   'app/lib/core/services/app_event_handler/app_event_handler.dart',
   'app/lib/core/services/app_event_handler/app_event_handler_navigation.dart',
   'app/lib/core/services/app_event_handler/app_event_handler_unit_panels.dart',
+  'app/lib/core/services/game_service/game_service.dart',
+  'app/lib/core/services/game_service/game_service_map_cache.dart',
+  'app/lib/core/services/game_service/game_service_new_game_setup.dart',
+  'app/lib/core/services/game_service/game_service_new_game_setup_maps.dart',
+  'app/lib/core/services/game_service/game_service_new_game_setup_pipeline.dart',
+  'app/lib/core/services/game_service/game_service_turn_resume.dart',
+  'app/lib/core/services/game_service/game_service_turn_trace.dart',
+  'app/lib/core/services/game_service/game_service_turn_trace_ai_sections.dart',
   'app/lib/core/services/debug/app_event_handler_debug_flip_province.dart',
   'app/lib/core/services/debug/app_event_handler_debug_flip_province_resolve.dart',
   'app/lib/core/services/debug/app_event_handler_debug_set_diplomacy.dart',
